+++ internal/resp/fuzz_test.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/kanata996/hah/errx"
+	"github.com/kanata996/hah/internal/errx"
 )
 
 func FuzzJSONWritersPublicContracts(f *testing.F) {